--- src/main/java/io/urdego/urdego_game_service/api/game/dto/request/GameCreateReq.java
@@ -1,6 +1,0 @@
-package io.urdego.urdego_game_service.api.game.dto.request;
-
-public record GameCreateReq(
-    String roomId
-) {
-}
