--- src/main/java/io/urdego/urdego_game_service/api/round/dto/request/QuestionReq.java
@@ -1,6 +1,0 @@
-package io.urdego.urdego_game_service.api.round.dto.request;
-
-public record QuestionReq(
-        String questionId
-) {
-}
