+++ src/main/java/io/urdego/urdego_game_service/controller/game/dto/request/GameCreateReq.java
@@ -1,0 +1,6 @@
+package io.urdego.urdego_game_service.controller.game.dto.request;
+
+public record GameCreateReq(
+    String roomId
+) {
+}
