+++ src/main/java/io/urdego/urdego_game_service/controller/room/dto/request/ContentSelectReq.java
@@ -1,4 +1,4 @@
-package io.urdego.urdego_game_service.api.room.dto.request;
+package io.urdego.urdego_game_service.controller.room.dto.request;
 
 import java.util.List;
