+++ src/main/java/io/urdego/urdego_game_service/controller/room/dto/response/RoomInfoRes.java
@@ -1,17 +1,17 @@
-package io.urdego.urdego_game_service.api.room.dto.response;
+package io.urdego.urdego_game_service.controller.room.dto.response;
 
 import io.urdego.urdego_game_service.common.enums.Status;
 import io.urdego.urdego_game_service.domain.room.entity.Room;
 
 import java.util.List;
 
-public record RoomCreateRes(
+public record RoomInfoRes(
         String roomId,
         Status status,
         List<String> currentPlayers
 ) {
-    public static RoomCreateRes from(Room room) {
-        return new RoomCreateRes(
+    public static RoomInfoRes from(Room room) {
+        return new RoomInfoRes(
                 room.getRoomId(),
                 room.getStatus(),
                 room.getCurrentPlayers()
