+++ src/main/java/io/urdego/urdego_game_service/controller/round/dto/request/AnswerReq.java
@@ -1,4 +1,4 @@
-package io.urdego.urdego_game_service.api.round.dto.request;
+package io.urdego.urdego_game_service.controller.round.dto.request;
 
 public record AnswerReq(
         String questionId,
