+++ src/main/java/io/urdego/urdego_game_service/controller/round/dto/request/QuestionReq.java
@@ -1,0 +1,6 @@
+package io.urdego.urdego_game_service.controller.round.dto.request;
+
+public record QuestionReq(
+        String questionId
+) {
+}
