+++ src/main/java/io/urdego/urdego_game_service/domain/room/service/RoomService.java
@@ -1,14 +1,12 @@
 package io.urdego.urdego_game_service.domain.room.service;
 
-import io.urdego.urdego_game_service.api.room.dto.request.ContentSelectReq;
-import io.urdego.urdego_game_service.api.room.dto.request.RoomCreateReq;
-import io.urdego.urdego_game_service.api.room.dto.response.RoomCreateRes;
-import io.urdego.urdego_game_service.api.room.dto.response.RoomInfoRes;
+import io.urdego.urdego_game_service.controller.room.dto.request.ContentSelectReq;
+import io.urdego.urdego_game_service.controller.room.dto.request.RoomCreateReq;
+import io.urdego.urdego_game_service.controller.room.dto.response.RoomCreateRes;
+import io.urdego.urdego_game_service.controller.room.dto.response.RoomInfoRes;
 import io.urdego.urdego_game_service.common.enums.Status;
 import io.urdego.urdego_game_service.domain.room.entity.Room;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 public interface RoomService {
